Updated RecipeScreen and CategoryItem to navigate to detail screen
The `RecipeScreen` Composable now accepts a `navigateToDetail` lambda function, which is passed down to `CategoryScreen` and then to `CategoryItem`.

The `CategoryItem` Composable has been made clickable, and when clicked, it invokes the `navigateToDetail` function with the corresponding `Category` object. This allows navigation to a detail screen when a category item is clicked.